typecast the type to an int on return

+++ tests/libtest/lib500.c
@@ -8,6 +8,6 @@ int test(char *URL)
   curl_easy_setopt(curl, CURLOPT_HEADER, TRUE);
   res = curl_easy_perform(curl);
   curl_easy_cleanup(curl);  
-  return res;
+  return (int)res;
 }
